JDBC Controllers set

+++ src/main/java/com/itacademy/repository/impl/UserDaoImpl.java
@@ -5,13 +5,12 @@ import com.itacademy.domain.User;
 import com.itacademy.domain.VariantAnswer;
 import com.itacademy.repository.RoleDao;
 import com.itacademy.repository.UserDao;
-import com.itacademy.repository.VariantsAnswerDao;
+import com.itacademy.repository.VariantAnswerDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
@@ -19,7 +18,6 @@ import org.springframework.stereotype.Repository;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
@@ -114,15 +112,15 @@ public class UserDaoImpl implements UserDao {
 
 
     @Override
-    public Role setUserRole(User user, Role role) {
+    public Role setUserRole(Long userId, Long roleId) {
         final String saveUserQuery = "INSERT INTO users_roles(user_id, role_id) " +
                 "VALUES (:userId, :roleId)";
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         MapSqlParameterSource params = new MapSqlParameterSource();
-        params.addValue("userId", user.getUserId());
-        params.addValue("roleId", role.getRoleId());
+        params.addValue("userId", userId);
+        params.addValue("roleId", roleId);
         namedParameterJdbcTemplate.update(saveUserQuery, params, keyHolder);
 
         long createdRoleId = Objects.requireNonNull(keyHolder.getKey()).longValue();
@@ -130,52 +128,59 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public List<Role> getUserRoles(User user) {
+    public List<Role> getUserRoles(Long userId) {
         final String getUsersRolesQuery = "SELECT * from voting.users_roles ur left join voting.roles r on r.role_id = ur.role_id" +
                 " where user_id = :userId";
 
         MapSqlParameterSource params = new MapSqlParameterSource();
-        params.addValue("userId", user.getUserId());
+        params.addValue("userId", userId);
 
         return namedParameterJdbcTemplate.query(getUsersRolesQuery, params, roleDao::getRoleFromRow);
     }
 
     @Override
-    public void deleteUserRole(User user, Role role) {
-        final String deleteUserRole = "DELETE from users_roles where role_id = :roleId";
+    public void deleteUserRole(Long userId, Long roleId) {
+        final String deleteUserRole = "DELETE from voting.users_roles where user_id=:userId and role_id = :roleId";
         MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("roleId", role.getRoleId());
+        param.addValue("userId", userId);
+        param.addValue("roleId", roleId);
         namedParameterJdbcTemplate.update(deleteUserRole, param);
     }
 
 
     @Autowired
-    VariantsAnswerDao variantAnswerDao;
+    VariantAnswerDao variantAnswerDao;
 
     @Override
-    public void setUserVariantAnswer(User user, VariantAnswer variantAnswer) {
+    public void setUserVariantAnswer(Long userId, Long answerId) {
 
         final String saveUserVariantAnswer = "INSERT INTO voting.users_answers(answer_date, user_id, answer_id) " +
                 "VALUES (:answerDate, :userId, :variantAnswerId)";
 
         MapSqlParameterSource params = new MapSqlParameterSource();
         params.addValue("answerDate", LocalDateTime.now());
-        params.addValue("userId", user.getUserId());
-        params.addValue("variantAnswerId", variantAnswer.getAnswerId());
+        params.addValue("userId", userId);
+        params.addValue("variantAnswerId", answerId);
         namedParameterJdbcTemplate.update(saveUserVariantAnswer, params);
     }
 
     @Override
-    public List<VariantAnswer> getUserVariantsAnswer(User user) {
-        final String getUserVariantAnswer = "select * from voting.users_answers join voting.variant_answers on users_answers.answer_id = variant_answers.answer_id" +
+    public List<VariantAnswer> getUserVariantsAnswer(Long id) {
+        final String getUserVariantAnswer = "select * from voting.users_answers " +
+                "join voting.variant_answers on users_answers.answer_id = variant_answers.answer_id" +
                 " where user_id = :userId";
         MapSqlParameterSource params = new MapSqlParameterSource();
-        params.addValue("userId", user.getUserId());
+        params.addValue("userId", id);
         return namedParameterJdbcTemplate.query(getUserVariantAnswer,params, variantAnswerDao::getAnswerFromRow);
     }
 
     @Override
-    public void deleteVariantAnswer(User user, VariantAnswer variantAnswer) {
+    public void deleteVariantAnswer(Long userId, Long answerId) {
+        final String deleteVariantAnswer = "DELETE from voting.users_answers where user_id = :userId and answer_id = :answerId";
+        MapSqlParameterSource param = new MapSqlParameterSource();
+        param.addValue("userId", userId);
+        param.addValue("answerId", answerId);
+        namedParameterJdbcTemplate.update(deleteVariantAnswer, param);
 
     }
